requirements.txt
pandas
numpy
python-dotenv
mysql-connector-python
SQLAlchemy
pymysql